meta: archive repository
To anyone who might be reading this: I am sick and tired of constantly being
plagiarized by those audacious enough to claim my work is theirs. This is not
the spirit of open source. This is not the spirit of programming. What this
is, is a clear violation of the basic principles of respect and decency. At this
point I feel zero obligation to provide whatever limited service I could for the
NixOS community. You are not entitled to my time and efforts.

To all of those kind enough to give credit where it's due: a very big thank you.
I do not think I can convey how much I appreciate you. Some of you are the reason
this repository has been maintainned with special care. To the rest of you pricks
who think code being in the open is license to copy: you suck. I would hope that
you experience the same to understand how it feels, but I really doubt you have ever
produced a single original thought. So, may you all have the day you deserve.

I assume the exact person who has done the most copying will also be reading
this out of sheer curiosity, or due to the fact that they continue tracking this
repository, so here it is: **fuck you**. I genuinely tried talking to you like
a human being to make a basic request and you, without a speck of dignity, chose
to be a prick. I will not indulge you further. Write your own code, or fuck right
off. You cannot become better at what you do by copying my sense of humor ***and***
my code. I expect the former is harder for someone such as yourself.

This configuration has been abandoned. It will not receive any further
updates. If you were tracking this repository for some reason, I suggest
that you find a hobby as soon as possible. What the hell were you thinking?
Why would you do such a thing to yourself? Are you illiterate perhaps?

P.S. I am *still open to civilized discussions* regarding Nix or NixOS over any
of the channels mentioned in my GitHub profile. If you are confused and overwhelmed
by Nix/NixOS, or perhaps got somewhere, please do feel free to contact me. I am and
will remain open to discussing Nix/NixOS in a context in which I am not constantly
plagiarized by children pretending to be programmers.

Farewell.